Family Pack chega ao loop de render: IFile/IGL/EGL corrigidos contra o layout real
Sessao de depuracao dirigida por smoke test contra a ROM real do Zeebo
Family Pack. Estado antes: CPU descarrilava em 0xE92D00F0 com SP=0 durante
o CreateInstance. Estado depois: boot completo -> EVT_APP_START ->
"rodando", pilha EGL inicializada de ponta a ponta, data.vfs aberto e
streamado (header FUFS + indice + chunks 4KB), e glClear real chegando
pela vtable e pintando o framebuffer (fb=0xFFFFFFFF).

Causas raiz encontradas e corrigidas (todas confirmadas contra o codigo
do zeemu, que roda jogos comerciais):

- ifile.c: vtable do IFile fora de ordem (sem Readable/Cancel do
  IAStream) fazia IFILE_Read do jogo cair no trap de WRITE - era isso que
  gravava 4 bytes zero no offset 0 do data.vfs (magic FUFS destruido, ja
  restaurado do git). Modo de abertura tratava _OFM_READ (so-leitura)
  como gravavel. Seek usava ordem stdio (1=atual 2=fim) em vez da ordem
  AEE real (1=FIM 2=ATUAL). Vtables IFile (12 slots) e IFileMgr (21
  slots) reescritas no layout real do AEEFile.h; modo OFM como bitmask;
  GetInfo/Readable/Cancel/GetPos implementados.

- boot.c: rotear AEECLSID_DISPLAY_REAL para vtable propria de 10 slots em
  ordem inventada fazia o jogo ler alem do fim dela (GetDeviceBitmap e o
  slot 16 do layout real de 48), saltar para 0x0, re-executar o entry do
  modulo (zerando o BSS todo) e descarrilar com SP=0 - era o crash do
  0xE92D00F0. Revertido para o stub generico consciente de clsid;
  GetFontMetrics/MeasureTextEx (slots 2/3 reais) preenchem metricas.

- egl_gl.c: vtable IGL renumerada para o layout real (COM 0-2 + funcoes
  GLES 1.x em ordem alfabetica a partir do slot 3, ~80 slots, igual ao
  BrewGL do zeemu) - antes glClear/GenTextures/etc estavam em slots
  inventados que nao batiam com nada, e o "GenTextures" no slot 63 (na
  verdade glRotatex) podia fazer milhoes de writes com o angulo como
  contador. Tabela direta "sem this" espelhada no corpo do objeto.
  eglGetConfigAttrib devolve valores reais (RGB565: BUFFER_SIZE=16 etc -
  zerar isso virava divisor zero no calculo de bytes-por-pixel do jogo).
  eglSwapBuffers apresenta em vez de apagar o framebuffer. glClearColorx
  converte fixed-point 16.16 corretamente.

- helpers.c: religa wstrlen/strtowstr/wstrtostr removidos num refactor
  inacabado (as estaticas guest_* existiam sem uso); corrige capacidade
  dividida por 2 na guest_wstrtostr.

- mod_loader.c: rejeita magic PK/GZIP com erro claro em vez de executar
  o container como codigo ARM em VA 0.

- .gitignore: excecao !src/debug/ - o padrao [Dd]ebug/ de Visual Studio
  engolia o diretorio de CODIGO src/debug (log/trace), entao log.c/
  trace.c/disasm.c/log.h nunca tinham sido commitados e um clone do repo
  nao compilava. Agora estao no git.

Bloqueio restante (diagnosticado, nao resolvido): o jogo estagna num
busy-wait no callback de timer - divisao por subtracao com divisor zero
propagado de um ponteiro de tabela NULL num objeto do scheduler do
proprio jogo (NULL-deref le bytes do modulo em VA 0 sem falhar). Detalhes
e proximos passos em docs/PROGRESS.md. Watchpoints temporarios de debug
em cpu.c/memory.c/ifile.c documentados la.

+++ src/audio/mixer.c
@@ -27,10 +27,12 @@ void zaudio_render(int16_t *out, uint32_t frames) {
             s = (s * (int32_t)voice->volume) / 255;
 
             /* mono -> estereo, acumula com clamp */
-            int32_t l = out[f * 2]     + s;
-            int32_t r = out[f * 2 + 1] + s;
-            if (l > 32767) l = 32767; if (l < -32768) l = -32768;
-            if (r > 32767) r = 32767; if (r < -32768) r = -32768;
+            int32_t l = (int32_t)out[f * 2] + s;
+            int32_t r = (int32_t)out[f * 2 + 1] + s;
+            if (l > 32767) l = 32767;
+            if (l < -32768) l = -32768;
+            if (r > 32767) r = 32767;
+            if (r < -32768) r = -32768;
             out[f * 2]     = (int16_t)l;
             out[f * 2 + 1] = (int16_t)r;
